refactor: remove an unnecessary intermediate directory
