+++ src/core/apps/About.cpp
