+++ src/core/apps/About.h
