+++ src/core/apps/AppFunctions.h
