+++ src/core/apps/AppLauncher.cpp
