+++ src/core/apps/CMakeLists.txt
