+++ src/core/apps/NotesApp.cpp
