+++ src/core/apps/NotesApp.h
