+++ src/core/apps/components/CMakeLists.txt
