--- src/core/avitab/CMakeLists.txt
@@ -1,8 +1,0 @@
-# This file is part of the Avitab project. See the README and LICENSE for details.
-
-target_sources(avitab_core_lib PRIVATE
-    AviTab.cpp
-    AviTab.h
-)
-
-add_subdirectory(apps)
